Fixes for qt in Python 3.10 from electron-cash (#10)
* Fixup for Python 3.10

This commit contains some fixups for Python 3.10.  Apparently some of
the PyQt overloads disappeared that support int <-> float
interchangeably as args to C++ functions, which caused the app to not
work at all on Python 3.10.

Note: there are still extant bugs with using Python 3.10, but this
commit at least allows the app to (mostly) run ok.

* More fixups for Python 3.10

For some reason now, PyQt5 doesn't auto-coerce args from int -> float
and float -> int, so everything that is passing the wrong type from
Python to PyQt is broken now.  Luckily this doesn't happen in too many
places, but it does happen.

This commit fixes more of these errors leading to app crash.

* Python 3.10 fixups

Fixed a bunch of places where potentially non-bool or non-int params are
being passed to functions expecting precisely bools and/or ints.

Also fixed a typo in main_window.py.

Co-authored-by: Calin Culianu <calin.culianu@gmail.com>