Tor: Fix closing of stdout on Windows
Due to the `close_output` argument in stem's `launch_tor_with_config`
not being passed to `launch_tor` the Tor process' `stdout` is closed
which makes the Tor logging not work.

We monkey patch `launch_tor` to always set `close_output` to `False`.

This will not be needed anymore when a new stem version that includes
this commit is released:
https://github.com/torproject/stem/commit/7a4357b2c21d0af5088c1cafce0800fc63cebbb4